chore(repeated-string-transformation): upd solutions
Though the third solution is correct the explanation is not; the claim was that characters would be in the same connected component iff they resolved to the same character.

However, this is not the case, as shown by {a -> c} - a and c both resolve to c, but {a, c} is not a connected component -- it does not satisfy the property of reflexivity.

Though the solution can still be explained without it, it's not nearly as easy to do so without being confusing. So, just remove it for now.